EntityTeleport refactor to v4567
